Remove old vector.rs
