+++ ext/crates/fp/Cargo.toml
@@ -9,8 +9,6 @@ edition = "2018"
 [dependencies]
 serde_json = { version = "1.0.0", optional = true }
 serde = { version = "1.0.0", optional = true }
-# Cannot upgrade due to https://gitlab.com/antonok/enum_dispatch/-/issues/34
-enum_dispatch = "0.2.0"
 saveload = { path = "../saveload" }
 itertools = "0.9.0"
 syn = "=1.0.57"
