+++ ext/crates/fp/src/lib.rs
@@ -1,10 +1,8 @@
-#![cfg_attr(rustfmt, rustfmt_skip)]
 #![feature(const_panic)]
 #![allow(clippy::many_single_char_names)]
 #![allow(clippy::unreadable_literal)]
 
+pub mod matrix;
 pub mod prime;
 pub mod vector;
-pub mod matrix;
-pub mod vector_new;
 mod vector_inner;
